Move handleClips out of the server action surface
handleClips mutates clip and lesson balances and was exported from a
"use server" file, making it an endpoint anyone could call. It was not
exploitable — its first argument is a Prisma transaction client, which
cannot be serialised across the wire, so an outside call dies on
tx.purchaseItem — but it had no business being reachable at all.

It now lives in lib/clips.ts, a plain server module with no "use server",
alongside the PrismaTx type it needs. The four callers (admin, orders,
admin-terminer, server-actions) import it from there. purchase-actions.ts
keeps "use server" because getProductStats is genuinely called from
client components.
